+++ resources/views/sections/CARTDT01001/section.blade.php
@@ -87,7 +87,7 @@
 				@endif
 				
 				<?php
-					$services = \App\CartService::getAvailableCartServices();
+					$services = \TheRealJanJanssens\Pakka\Models\CartService::getAvailableCartServices();
 					//dd($services);
 				?>
 				
@@ -117,8 +117,8 @@
 				
 				<div class="boxed e pb-0 {{ parseSecAttr('.e', $section['classes']) }}">
 	                <?php
-		                $regions = App\ShipmentOption::getAvailableRegions();
-		                $shipment_options = App\ShipmentOption::getAvailableOptions();
+		                $regions = TheRealJanJanssens\Pakka\Models\ShipmentOption::getAvailableRegions();
+		                $shipment_options = TheRealJanJanssens\Pakka\Models\ShipmentOption::getAvailableOptions();
 		                
 		                $shipment_id = null;
 		                if(Cart::getCondition('SHIPPING')){