+++ resources/views/sections/CTACHR01001/section.blade.php
@@ -1,0 +1,22 @@
+<section id="{{ $section['id'] }}" class="space--xs {{ checkAdjustable() }} {{ checkManageable() }} {{ parseSecAttr('.adjustable', $section['classes']) }}" {{ parseEditSecAttr($page['meta']['mode'],$section) }} {{ parseSecAttr('.adjustable', $section['attributes']) }}>
+    <div class="container">
+        <div class="cta cta--horizontal text-center-xs row">
+            <div class="col-md-4">
+                <h4>{{ parseContent($section['CTACHR01001_H'],'title') }}</h4>
+            </div>
+            <div class="col-md-5">
+                <p class="lead">{{ parseContent($section['CTACHR01001_H'],'text') }}</p>
+            </div>
+            <div class="col-md-3 text-right text-center-xs">
+                @if(checkContent($section['CTACHR01001_H'], 'link'))
+	                <a class="btn btn--primary type--uppercase e {{ parseSecAttr('.e', $section['classes']) }}" href="{{ $section['CTACHR01001_H']['link'] }}">
+		                <span class="btn__text">
+							{{ parseContent($section['CTACHR01001_H'],'button') }}
+						</span> 
+					</a>
+	            @endif
+            </div>
+        </div>
+    </div>
+{!! constructDividers($section) !!}
+</section>