+++ src/Helpers/helpers.php
@@ -2849,7 +2849,7 @@ if (! function_exists('move_file')) {
         $destinationPath = config('image.'.$type.'.folder');
         $width = config('image.' . $type . '.width');
         $height = config('image.' . $type . '.height');
-        $full_name = str_random(16) . '.' . $file->getClientOriginalExtension();
+        $full_name = generateString(16) . '.' . $file->getClientOriginalExtension();
         
         if ($width == null && $height == null) { // Just move the file
             $file->storeAs($destinationPath, $full_name);